Web: Define TabPage, use it for search, song, genre and label page
