+++ front/apps/web/src/pages/genres/[slugOrId].tsx
@@ -16,18 +16,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-	Box,
-	IconButton,
-	Skeleton,
-	Tab,
-	Tabs,
-	Typography,
-} from "@mui/material";
+import { Box, IconButton, Skeleton, Typography } from "@mui/material";
 import { useSetAtom } from "jotai";
 import type { NextPageContext } from "next";
 import { useRouter } from "next/router";
-import { useTranslation } from "react-i18next";
 import type { GetPropsTypesFrom, Page } from "ssr";
 import { getAlbums, getArtists, getGenre, getSongs } from "@/api/queries";
 import {
@@ -41,7 +33,7 @@ import { Head } from "~/components/head";
 import InfiniteAlbumView from "~/components/infinite/resource/album";
 import InfiniteArtistView from "~/components/infinite/resource/artist";
 import { InfiniteSongView } from "~/components/infinite/resource/song";
-import { useTabRouter } from "~/components/tab-router";
+import { TabPage } from "~/components/tab-page";
 import getSlugOrId from "~/utils/getSlugOrId";
 
 const prepareSSR = (context: NextPageContext) => {
@@ -73,16 +65,8 @@ const tabs = ["artist", "album", "song"] as const;
 
 const GenrePage: Page<GetPropsTypesFrom<typeof prepareSSR>> = ({ props }) => {
 	const router = useRouter();
-	const { t } = useTranslation();
 	const genreIdentifier = props?.genreIdentifier ?? getSlugOrId(router.query);
 	const genre = useQuery(getGenre, genreIdentifier);
-	const { selectedTab, selectTab } = useTabRouter(
-		(r) => r.query.t,
-		(newTab) => `/genres/${genreIdentifier}?t=${newTab}`,
-		"album",
-		"artist",
-		"song",
-	);
 	const playFromQuery = useSetAtom(playFromInfiniteQuery);
 	const queryClient = useQueryClient();
 	const playRadio = () => {
@@ -99,87 +83,97 @@ const GenrePage: Page<GetPropsTypesFrom<typeof prepareSSR>> = ({ props }) => {
 	};
 
 	return (
-		<Box sx={{ width: "100%" }}>
-			<Head title={genre.data?.name} />
-			<Box
-				sx={{
-					width: "100%",
-					gap: 1,
-					justifyContent: "center",
-					alignItems: "center",
-					textAlign: "center",
-					display: "flex",
-					marginY: 5,
-				}}
-			>
-				<IconButton onClick={playRadio}>
-					<RadioIcon />
-				</IconButton>
-				<Typography variant="h5" sx={{ fontWeight: "bold" }}>
-					{genre.data?.name ?? <Skeleton width={"100px"} />}
-				</Typography>
-			</Box>
-			<Tabs
-				value={selectedTab}
-				onChange={(__, tabName) => selectTab(tabName)}
-				variant="fullWidth"
-			>
-				{tabs.map((value, index) => (
-					<Tab
-						key={index}
-						value={value}
-						sx={{ minWidth: "fit-content", flex: 1 }}
-						label={t(`models.${value}_plural`)}
-					/>
-				))}
-			</Tabs>
-			<Box sx={{ paddingBottom: 2 }} />
-			{selectedTab === "artist" && (
-				<InfiniteArtistView
-					query={({ libraries, sortBy, order }) =>
-						getArtists(
-							{
-								genre: genreIdentifier,
-								library: libraries,
-							},
-							{ sortBy, order },
-							["illustration"],
-						)
-					}
-				/>
-			)}
-			{selectedTab === "album" && (
-				<InfiniteAlbumView
-					query={({ libraries, types, sortBy, order }) =>
-						getAlbums(
-							{
-								genre: genreIdentifier,
-								type: types,
-								library: libraries,
-							},
-							{ sortBy, order },
-							["artists", "illustration"],
-						)
-					}
-				/>
-			)}
-			{selectedTab === "song" && (
-				<InfiniteSongView
-					query={({ libraries, types, random, sortBy, order }) =>
-						getSongs(
-							{
-								genre: genreIdentifier,
-								type: types,
-								random,
-								library: libraries,
-							},
-							{ sortBy, order },
-							["artist", "featuring", "master", "illustration"],
-						)
-					}
-				/>
-			)}
-		</Box>
+		<TabPage
+			tabs={tabs}
+			urlFromTab={(newTab) => `/genres/${genreIdentifier}?t=${newTab}`}
+			translateTab={(tab) => `models.${tab}_plural`}
+			header={
+				<>
+					<Head title={genre.data?.name} />
+					<Box
+						sx={{
+							width: "100%",
+							gap: 1,
+							justifyContent: "center",
+							alignItems: "center",
+							textAlign: "center",
+							display: "flex",
+							marginY: 5,
+						}}
+					>
+						<IconButton onClick={playRadio}>
+							<RadioIcon />
+						</IconButton>
+						<Typography variant="h5" sx={{ fontWeight: "bold" }}>
+							{genre.data?.name ?? <Skeleton width={"100px"} />}
+						</Typography>
+					</Box>
+				</>
+			}
+			render={(tab) => {
+				switch (tab) {
+					case "artist":
+						return (
+							<InfiniteArtistView
+								query={({ libraries, sortBy, order }) =>
+									getArtists(
+										{
+											genre: genreIdentifier,
+											library: libraries,
+										},
+										{ sortBy, order },
+										["illustration"],
+									)
+								}
+							/>
+						);
+					case "album":
+						return (
+							<InfiniteAlbumView
+								query={({ libraries, types, sortBy, order }) =>
+									getAlbums(
+										{
+											genre: genreIdentifier,
+											type: types,
+											library: libraries,
+										},
+										{ sortBy, order },
+										["artists", "illustration"],
+									)
+								}
+							/>
+						);
+					case "song":
+						return (
+							<InfiniteSongView
+								query={({
+									libraries,
+									types,
+									random,
+									sortBy,
+									order,
+								}) =>
+									getSongs(
+										{
+											genre: genreIdentifier,
+											type: types,
+											random,
+											library: libraries,
+										},
+										{ sortBy, order },
+										[
+											"artist",
+											"featuring",
+											"master",
+											"illustration",
+										],
+									)
+								}
+							/>
+						);
+				}
+			}}
+		/>
 	);
 };
