+++ front/apps/web/src/pages/songs/[slugOrId]/[...tab].tsx
@@ -16,16 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-	Box,
-	Button,
-	Divider,
-	Skeleton,
-	Stack,
-	Tab,
-	Tabs,
-	Typography,
-} from "@mui/material";
+import { Button, Divider, Skeleton, Stack, Typography } from "@mui/material";
 import type { QueryClient } from "@tanstack/react-query";
 import { useSetAtom } from "jotai";
 import type { NextPageContext } from "next";
@@ -57,7 +48,7 @@ import InfiniteVideoView from "~/components/infinite/resource/video";
 import LyricsBox from "~/components/lyrics";
 import SongRelationPageHeader from "~/components/relation-page-header/resource/song";
 import SongTypeIcon from "~/components/song-type-icon";
-import { useTabRouter } from "~/components/tab-router";
+import { TabPage } from "~/components/tab-page";
 import getSlugOrId from "~/utils/getSlugOrId";
 import { useThemedSxValue } from "~/utils/themed-sx-value";
 
@@ -121,11 +112,6 @@ const getTabLabel = (tab: (typeof tabs)[number]): TranslationKey => {
 };
 
 const SongPage: Page<GetPropsTypesFrom<typeof prepareSSR>> = ({ props }) => {
-	const { selectedTab, selectTab } = useTabRouter(
-		(router) => router.query.tab,
-		(newTab) => `/songs/${songIdentifier}/${newTab}`,
-		...tabs,
-	);
 	const { t } = useTranslation();
 	const router = useRouter();
 	const playTrack = useSetAtom(playTrackAtom);
@@ -153,255 +139,285 @@ const SongPage: Page<GetPropsTypesFrom<typeof prepareSSR>> = ({ props }) => {
 
 	return (
 		<>
-			<Head title={song.data?.name} />
-			<GradientBackground />
-			<SongRelationPageHeader song={song.data} />
-			<Button
-				variant="contained"
-				sx={{ width: "100%", marginTop: 1 }}
-				endIcon={<PlayIcon />}
-				onClick={
-					song.data &&
-					(() =>
-						playTrack({
-							track: {
-								...song.data.master,
-								illustration: song.data.illustration,
-							},
-							artist: song.data.artist,
-							featuring: song.data.featuring,
-						}))
-				}
-			>
-				{t("actions.playback.play")}
-			</Button>
-			<Divider sx={{ paddingY: 1 }} />
-			<Tabs
-				value={selectedTab}
-				onChange={(__, tabName) => selectTab(tabName)}
-				variant="scrollable"
-			>
-				{tabs.map((value, index) => (
-					<Tab
-						key={index}
-						value={value}
-						sx={{ minWidth: "fit-content", flex: 1 }}
-						label={t(getTabLabel(value))}
-					/>
-				))}
-			</Tabs>
-			<Box sx={{ paddingY: 2 }}>
-				{selectedTab === "info" && (
+			<TabPage
+				tabs={tabs}
+				getTabFromRouter={(r) => r.query.tab}
+				urlFromTab={(tab) => `/songs/${songIdentifier}/${tab}`}
+				translateTab={getTabLabel}
+				header={
 					<>
-						<Head
-							title={
+						<Head title={song.data?.name} />
+						<GradientBackground />
+						<SongRelationPageHeader song={song.data} />
+						<Button
+							variant="contained"
+							sx={{ width: "100%", marginTop: 1 }}
+							endIcon={<PlayIcon />}
+							onClick={
 								song.data &&
-								`${song.data?.name} (${t("song.infoTab")})`
+								(() =>
+									playTrack({
+										track: {
+											...song.data.master,
+											illustration:
+												song.data.illustration,
+										},
+										artist: song.data.artist,
+										featuring: song.data.featuring,
+									}))
 							}
-						/>
-						<Stack
-							direction="row"
-							sx={{
-								overflowX: "scroll",
-								alignItems: "center",
-								paddingBottom: 2,
-							}}
-							spacing={1}
 						>
-							<Typography sx={{ overflow: "unset" }}>
-								{`${t("song.songType")}: `}
-							</Typography>
-							{song.data ? (
-								song.data.type ? (
+							{t("actions.playback.play")}
+						</Button>
+						<Divider sx={{ paddingY: 1 }} />
+					</>
+				}
+				render={(tab) => {
+					switch (tab) {
+						case "lyrics":
+							return (
+								<>
+									<Head
+										title={
+											song.data &&
+											`${song.data?.name} (${t("models.lyrics")})`
+										}
+									/>
+									<LyricsBox
+										songName={song.data?.name}
+										lyrics={
+											song.data
+												? (song.data.lyrics?.plain.split(
+														"\n",
+													) ?? null)
+												: undefined
+										}
+									/>
+								</>
+							);
+						case "versions":
+							return (
+								<>
+									<Head
+										title={
+											song.data &&
+											`${song.data.name} (${t("models.versions")})`
+										}
+									/>
+									<InfiniteSongView
+										disableShuffle
+										query={({
+											libraries,
+											sortBy,
+											order,
+											types,
+										}) =>
+											getSongs(
+												{
+													library: libraries,
+													type: types,
+													versionsOf: songIdentifier,
+												},
+												{ sortBy, order },
+												[
+													"artist",
+													"featuring",
+													"master",
+													"illustration",
+												],
+											)
+										}
+									/>
+								</>
+							);
+						case "videos":
+							return (
+								<>
+									<Head
+										title={
+											song.data &&
+											`${song.data.name} (${t("models.video_plural")})`
+										}
+									/>
+									<InfiniteVideoView
+										query={({
+											libraries,
+											sortBy,
+											order,
+											types,
+										}) =>
+											getVideos(
+												{
+													library: libraries,
+													type: types,
+													group: song.data?.groupId,
+												},
+												{ sortBy, order },
+												[
+													"artist",
+													"master",
+													"illustration",
+												],
+											)
+										}
+										subtitle="duration"
+									/>
+								</>
+							);
+						case "tracks":
+							return (
+								<>
+									<Head
+										title={
+											song.data &&
+											`${song.data.name} (${t("models.track_plural")})`
+										}
+									/>
+									<InfiniteTrackView
+										query={({ sortBy, order, libraries }) =>
+											getTracks(
+												{
+													song: songIdentifier,
+													library: libraries,
+												},
+												{ sortBy, order },
+												[
+													"release",
+													"song",
+													"illustration",
+													"video",
+												],
+											)
+										}
+									/>
+								</>
+							);
+						case "info":
+							return (
+								<>
+									<Head
+										title={
+											song.data &&
+											`${song.data?.name} (${t("song.infoTab")})`
+										}
+									/>
 									<Stack
 										direction="row"
+										sx={{
+											overflowX: "scroll",
+											alignItems: "center",
+											paddingBottom: 2,
+										}}
 										spacing={1}
-										sx={{ alignItems: "center" }}
 									>
-										<SongTypeIcon
-											type={song.data.type}
-											size={20}
-										/>
-										<Typography>
-											{t(
-												songTypeToTranslationKey(
-													song.data.type,
-													false,
-												),
-											)}
+										<Typography sx={{ overflow: "unset" }}>
+											{`${t("song.songType")}: `}
 										</Typography>
+										{song.data ? (
+											song.data.type ? (
+												<Stack
+													direction="row"
+													spacing={1}
+													sx={{
+														alignItems: "center",
+													}}
+												>
+													<SongTypeIcon
+														type={song.data.type}
+														size={20}
+													/>
+													<Typography>
+														{t(
+															songTypeToTranslationKey(
+																song.data.type,
+																false,
+															),
+														)}
+													</Typography>
+												</Stack>
+											) : null
+										) : (
+											<Skeleton width="100px" />
+										)}
 									</Stack>
-								) : null
-							) : (
-								<Skeleton width="100px" />
-							)}
-						</Stack>
-						{song.data ? (
-							song.data.bpm ? (
-								<Typography
-									sx={{ paddingTop: 1, paddingBottom: 4 }}
-								>{`BPM: ${song.data.bpm}`}</Typography>
-							) : null
-						) : (
-							<Skeleton width={"50px"} />
-						)}
-						{(genres.items === undefined ||
-							genres.items.length !== 0) && (
-							<Stack
-								direction="row"
-								sx={{
-									overflowX: "scroll",
-									alignItems: "center",
-								}}
-								spacing={2}
-							>
-								<Typography sx={{ overflow: "unset" }}>
-									{`${t("models.genre_plural")}: `}
-								</Typography>
-								{(genres.items ?? generateArray(2)).map(
-									(genre, index) => (
-										<GenreChip
-											variant="outlined"
-											sx={genreButtonOutline}
-											key={index}
-											genre={genre}
-										/>
-									),
-								)}
-							</Stack>
-						)}
-						{song.data === undefined ||
-						externalMetadata.data?.sources.length ? (
-							<Stack
-								direction="row"
-								sx={{
-									overflowX: "scroll",
-									alignItems: "center",
-									paddingTop: 2,
-								}}
-								spacing={2}
-							>
-								<Typography sx={{ overflow: "unset" }}>
-									{`${t("models.externalLink_plural")}: `}
-								</Typography>
-								{(
-									externalMetadata.data?.sources.filter(
-										({ url }) => url !== null,
-									) ?? generateArray(2)
-								).map((source, index) => (
-									<ExternalMetadataBadge
-										key={index}
-										source={source}
-									/>
-								))}
-							</Stack>
-						) : null}
-						<Typography variant="body1" sx={{ paddingTop: 2 }}>
-							{externalMetadata.data?.description}
-						</Typography>
-					</>
-				)}
-				{selectedTab === "lyrics" && (
-					<>
-						<Head
-							title={
-								song.data &&
-								`${song.data?.name} (${t("models.lyrics")})`
-							}
-						/>
-						<LyricsBox
-							songName={song.data?.name}
-							lyrics={
-								song.data
-									? (song.data.lyrics?.plain.split("\n") ??
-										null)
-									: undefined
-							}
-						/>
-					</>
-				)}
-				{selectedTab === "versions" && (
-					<>
-						<Head
-							title={
-								song.data &&
-								`${song.data.name} (${t("models.versions")})`
-							}
-						/>
-						<InfiniteSongView
-							disableShuffle
-							query={({ libraries, sortBy, order, types }) =>
-								getSongs(
-									{
-										library: libraries,
-										type: types,
-										versionsOf: songIdentifier,
-									},
-									{ sortBy, order },
-									[
-										"artist",
-										"featuring",
-										"master",
-										"illustration",
-									],
-								)
-							}
-						/>
-					</>
-				)}
-
-				{selectedTab === "videos" && (
-					<>
-						<Head
-							title={
-								song.data &&
-								`${song.data.name} (${t("models.video_plural")})`
-							}
-						/>
-						<InfiniteVideoView
-							query={({ libraries, sortBy, order, types }) =>
-								getVideos(
-									{
-										library: libraries,
-										type: types,
-										group: song.data?.groupId,
-									},
-									{ sortBy, order },
-									["artist", "master", "illustration"],
-								)
-							}
-							subtitle="duration"
-						/>
-					</>
-				)}
-				{selectedTab === "tracks" && (
-					<>
-						<Head
-							title={
-								song.data &&
-								`${song.data.name} (${t("models.track_plural")})`
-							}
-						/>
-						<InfiniteTrackView
-							query={({ sortBy, order, libraries }) =>
-								getTracks(
-									{
-										song: songIdentifier,
-										library: libraries,
-									},
-									{ sortBy, order },
-									[
-										"release",
-										"song",
-										"illustration",
-										"video",
-									],
-								)
-							}
-						/>
-					</>
-				)}
-			</Box>
+									{song.data ? (
+										song.data.bpm ? (
+											<Typography
+												sx={{
+													paddingTop: 1,
+													paddingBottom: 4,
+												}}
+											>{`BPM: ${song.data.bpm}`}</Typography>
+										) : null
+									) : (
+										<Skeleton width={"50px"} />
+									)}
+									{(genres.items === undefined ||
+										genres.items.length !== 0) && (
+										<Stack
+											direction="row"
+											sx={{
+												overflowX: "scroll",
+												alignItems: "center",
+											}}
+											spacing={2}
+										>
+											<Typography
+												sx={{ overflow: "unset" }}
+											>
+												{`${t("models.genre_plural")}: `}
+											</Typography>
+											{(
+												genres.items ?? generateArray(2)
+											).map((genre, index) => (
+												<GenreChip
+													variant="outlined"
+													sx={genreButtonOutline}
+													key={index}
+													genre={genre}
+												/>
+											))}
+										</Stack>
+									)}
+									{song.data === undefined ||
+									externalMetadata.data?.sources.length ? (
+										<Stack
+											direction="row"
+											sx={{
+												overflowX: "scroll",
+												alignItems: "center",
+												paddingTop: 2,
+											}}
+											spacing={2}
+										>
+											<Typography
+												sx={{ overflow: "unset" }}
+											>
+												{`${t("models.externalLink_plural")}: `}
+											</Typography>
+											{(
+												externalMetadata.data?.sources.filter(
+													({ url }) => url !== null,
+												) ?? generateArray(2)
+											).map((source, index) => (
+												<ExternalMetadataBadge
+													key={index}
+													source={source}
+												/>
+											))}
+										</Stack>
+									) : null}
+									<Typography
+										variant="body1"
+										sx={{ paddingTop: 2 }}
+									>
+										{externalMetadata.data?.description}
+									</Typography>
+								</>
+							);
+					}
+				}}
+			/>
 		</>
 	);
 };
